Fix a silly typo that was breaking the automatic close functionality for Wine Steam. Also, add advisory on how to make it work at all.
Basically, with current tools, there's no way to know what process was started by Wine Steam as a result of starting a game. So you have to tell Arcadia what the game binary is so that it can look for it, wait for it to close, and then shutdown Steam.