Add support for PCSX2 and Dolphin.
Dolphin is very straightforward but only because I hadn't started to do to it what I did to PCSX2.

PCSX2's runner lets you modify the emulator configuration based on the needs of your game or your session. For example, if you're on battery, you probably want to turn off all the shiny bells and whistles. Or maybe every game except one runs great with 3x internal res. You can set up such things here.

The main reason Dolphin *doesn't* have this is because I have realized that this project is over scope to get away with keeping it in Bash. It's time to move on to another language. So instead of adding the config processing functionality to Dolphin's runner, I'll just be rewriting Arcadia.